Add files via upload
This is a 16 potentiometer assembly pcb project. It has 16 signal pins, pozitif and gnd pins located left and right side. Both side is connected so there is no differentes between left and right pins. I designed like that becaose of if needed pins on left or right. The gerber has pozitif copper area on top layer and gnd copper area on bottom layer